pyproject.toml
[project]
name = "dnd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs>=25.4.0",
    "json5>=0.12.1",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pyinstaller>=6.16.0",
    "pyside6>=6.10.0",
    "websockets>=15.0.1",
]
